Refactor FilterNotNullOptionalExtensions
