+++ src/core-taggeds-optional/Optional/Optional.T.Extensions.FilterNotNull/FilterNotNull.Inner/InnerFilterNotNull.cs
@@ -1,0 +1,21 @@
+﻿using System.Runtime.CompilerServices;
+
+namespace System;
+
+partial class FilterNotNullOptionalExtensions
+{
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static Optional<T> InnerFilterNotNull<T>(this Optional<T?> optional)
+        =>
+        optional
+        .Filter(value => value is not null)
+        .Map(InnerToNonNullable);
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static Optional<T> InnerFilterNotNull<T>(this Optional<T?> optional)
+        where T : struct
+        =>
+        optional
+        .Filter(value => value is not null)
+        .Map(InnerToNonNullable);
+}
