+++ src/core-taggeds-optional/Optional/Optional.T.Extensions.FilterNotNull/FilterNotNull/FilterNotNull.cs
@@ -4,10 +4,10 @@ partial class FilterNotNullOptionalExtensions
 {
     public static Optional<T> FilterNotNull<T>(this Optional<T?> optional)
         =>
-        optional.Filter(value => value is not null).Map(InnerToNonNullable);
+        optional.InnerFilterNotNull();
 
     public static Optional<T> FilterNotNull<T>(this Optional<T?> optional)
         where T : struct
         =>
-        optional.Filter(value => value is not null).Map(InnerToNonNullable);
+        optional.InnerFilterNotNull();
 }
